Move files out from pages folder
Make showEdit flag optional.
